Implement basic use of serde over QUIC
Implement a basic use case of serde over the QUIC protocol. The client creates
a `Row` struct and serialises it using serde/bincode before sending it over the
connection stream.

The server then reads a message and deserialises the struct using serde/bincode
again, before displaying the struct and responding to the client.