Implement iteration and collection over QUIC
Implement the ability for the client to send much more data using a loop and
for the server to receive it all and parse it correctly.

This approach is still limited, as the server will buffer all data until the
stream is dropped, meaning that memory usage is high until the entire vector of
data has been sent.