Implement better streaming approach
Implement a better streaming approach using a `Buffer<T>` structure to store
the incoming bytes. This then provides an implementation of the `Iterator`
trait, which returns deserialised structs if the buffer has enough data. This
allows data to be processed as early as possible, instead of waiting for the
end of the stream.